Silence signed / unsigned comparison warning.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_4_BRANCH@28564 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ contrib/src/xrc/xh_statbar.cpp
@@ -3,7 +3,7 @@
 // Purpose:     XRC resource for wxStatusBar
 // Author:      Brian Ravnsgaard Riis
 // Created:     2004/01/21
-// RCS-ID:      
+// RCS-ID:      $Id$
 // Copyright:   (c) 2004 Brian Ravnsgaard Riis
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
@@ -27,7 +27,7 @@
 #include "wx/xrc/xh_statbar.h"
 #include "wx/statusbr.h"
 
-wxStatusBarXmlHandler::wxStatusBarXmlHandler() : 
+wxStatusBarXmlHandler::wxStatusBarXmlHandler() :
         wxXmlResourceHandler()
 {
     XRC_ADD_STYLE(wxST_SIZEGRIP);
@@ -49,8 +49,8 @@ wxObject *wxStatusBarXmlHandler::DoCreateResource()
     if (fields > 1 && !widths.IsEmpty())
     {
         int *width = new int[fields];
-  
-        for (unsigned int i = 0; i < fields; ++i)
+
+        for (unsigned int i = 0; i < (unsigned)fields; ++i)
         {
             width[i] = wxAtoi(widths.BeforeFirst(wxT(',')));
             if(widths.Find(wxT(',')))
